backend/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
requests==2.31.0
urllib3==2.2.0
Werkzeug==3.0.1
